refactor : Long to Integer
https://github.com/softeerbootcamp/Team3/issues/39

+++ backend/gathergo/src/main/java/lightning/gathergo/repository/ArticleRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ArticleRepository extends CrudRepository<Article, Long> {
+public interface ArticleRepository extends CrudRepository<Article, Integer> {
 
     // id, hostId, title, thumbnail, curr,
     // total, isClosed, content, meetingDay, location,
@@ -37,7 +37,7 @@ public interface ArticleRepository extends CrudRepository<Article, Long> {
     @Modifying
     @Query("insert into article (hostId, title, thumbnail, curr, total, isClosed, content, meetingDay, location, regionId, categoryId, uuid) " +
             "values (:hostId, :title, :thumbnail, :curr, :total, :isClosed, :content, :meetingDay, :location, :regionId, :categoryId, :uuid);")
-    public void save(Long hostId, String title, String thumbnail, int curr,
+    public void save(Integer hostId, String title, String thumbnail, int curr,
                          int total, boolean isClosed, String content, Timestamp meetingDay, String location, int regionId, int categoryId, String uuid);
 
     @Query(value = "select id from article order by id desc LIMIT 1")
@@ -72,9 +72,9 @@ public interface ArticleRepository extends CrudRepository<Article, Long> {
             "regionId = :regionId, categoryId = :categoryId where id=:id")
     public void updateArticleById(String title, String thumbnail, int curr,
                                   int total, boolean isClosed, String content, Timestamp meetingDay, String location,
-                                  int regionId, int categoryId, Long id);
+                                  int regionId, int categoryId, Integer id);
 
     @Modifying
     @Query("delete from article where id=:id")
-    public void deleteById(Long id);
+    public void deleteById(Integer id);
 }
